+++ pkg/kubectl/cmd/auth/cani.go
@@ -36,10 +36,10 @@ import (
 	discovery "k8s.io/client-go/discovery"
 	authorizationv1client "k8s.io/client-go/kubernetes/typed/authorization/v1"
 	"k8s.io/kubectl/pkg/util/printers"
+	rbacutil "k8s.io/kubectl/pkg/util/rbac"
 	"k8s.io/kubectl/pkg/util/templates"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	describeutil "k8s.io/kubernetes/pkg/kubectl/describe/versioned"
-	rbacutil "k8s.io/kubernetes/pkg/kubectl/util/rbac"
 )
 
 // CanIOptions is the start of the data required to perform the operation.  As new fields are added, add them here instead of
