+++ pkg/kubectl/describe/versioned/describe.go
@@ -66,13 +66,13 @@ import (
 	"k8s.io/client-go/tools/reference"
 	"k8s.io/klog"
 	deploymentutil "k8s.io/kubectl/pkg/util/deployment"
+	"k8s.io/kubectl/pkg/util/rbac"
 	"k8s.io/kubernetes/pkg/kubectl/describe"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 	"k8s.io/kubernetes/pkg/kubectl/util/certificate"
 	"k8s.io/kubernetes/pkg/kubectl/util/event"
 	"k8s.io/kubernetes/pkg/kubectl/util/fieldpath"
 	"k8s.io/kubernetes/pkg/kubectl/util/qos"
-	"k8s.io/kubernetes/pkg/kubectl/util/rbac"
 	resourcehelper "k8s.io/kubernetes/pkg/kubectl/util/resource"
 	"k8s.io/kubernetes/pkg/kubectl/util/slice"
 	storageutil "k8s.io/kubernetes/pkg/kubectl/util/storage"
